+++ polling_stations/apps/data_importers/ems_importers.py
@@ -10,7 +10,6 @@ import os
 import tempfile
 
 import requests
-from addressbase.models import Address
 from data_finder.helpers import PostcodeError, geocode_point_only
 from data_importers.addresshelpers import (
     format_polling_station_address,
@@ -25,7 +24,6 @@ from data_importers.data_types import AddressList, StationSet
 from django.contrib.gis.geos import Point
 from django.core.exceptions import ObjectDoesNotExist
 from django.utils.text import slugify
-from uk_geo_utils.helpers import Postcode
 
 """
 We see a lot of CSVs exported from Xpress
@@ -82,9 +80,6 @@ class BaseXpressCsvImporter(BaseCsvStationsCsvAddressesImporter, metaclass=abc.A
             [getattr(record, field).strip() for field in self.station_address_fields]
         )
 
-    def get_station_postcode(self, record):
-        return getattr(record, self.station_postcode_field).strip()
-
     def geocode_from_postcode(self, record):
         if not self.allow_station_point_from_postcode:
             return None
@@ -97,29 +92,6 @@ class BaseXpressCsvImporter(BaseCsvStationsCsvAddressesImporter, metaclass=abc.A
             return location_data.centroid
         except PostcodeError:
             return None
-
-    def geocode_from_uprn(self, record):
-        uprn = getattr(record, self.station_uprn_field)
-        uprn = uprn.lstrip("0")
-        ab_rec = Address.objects.get(uprn=uprn)
-        ab_postcode = Postcode(ab_rec.postcode)
-        station_postcode = Postcode(self.get_station_postcode(record))
-        if ab_postcode != station_postcode:
-            ab_address = ab_rec.address
-            rec_address = self.get_station_address(record).replace(os.linesep, ", ")
-            station_id = getattr(record, self.station_id_field)
-            message = "\n".join(
-                [
-                    "Geocoding with UPRN. Station record postcode does not match addressbase postcode.",
-                    f"Station address: '{rec_address}, {station_postcode.with_space}' (id: {station_id})",
-                    f"Addressbase: '{ab_address}, {ab_postcode.with_space}'",
-                    "SUGGESTION:",
-                    f"        # '{rec_address}, {station_postcode.with_space}' (id: {station_id})",
-                    f"        if record.{self.station_id_field} == '{station_id}': record = record._replace({self.station_postcode_field}='{ab_postcode.with_space}')",
-                ]
-            )
-            self.logger.log_message(logging.WARNING, message + "\n")
-        return ab_rec.location
 
     def get_station_point(self, record):
         location = None
